Made a bunch of changes so that the Hero can have a current weapon

+++ Classes/Hero/Mage/mage.cpp
@@ -1,5 +1,4 @@
 #include "mage.h"
-#include <string>
 
 void Mage::setMagic(double newMagic) {
 	if (newMagic <= 0 || newMagic > maxMagic)
@@ -7,7 +6,7 @@ void Mage::setMagic(double newMagic) {
 	magic = newMagic;
 }
 
-Mage::Mage(string newName, double health, int lives, double magic): Hero(newName, health, lives) {
+Mage::Mage(string newName, int health, int lives, double magic): Hero(newName, health, lives) {
 	//name = newName;
 	//setHealth(health);
 	//setLevel(level);
